# Speedup find_clusters