--- a/pyuvdata/utils.py
+++ b/pyuvdata/utils.py
@@ -1148,7 +1148,10 @@ def find_clusters(location_ids, location_vectors, tol, strict=False):
         location vectors, can be multidimensional
     tol : float
         tolerance for clusters
-
+    strict : bool
+        Require that all adjacency lists be isolated maximal cliques.
+        This ensures that vectors do not fall into multiple clusters.
+        Default: False
     Returns
     -------
     list of list of location_ids