<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->

1. Break long baseline lists into chunks to avoid overrunning memory.
2. Structure adjacency as a list of sets to avoid doing type conversions inside loops.
3. Convert adjacency sets to frozensets to make them hashable.
4. Identify redundant baselines as those who all share the same adjacency set.
4. When baselines are added to redundant groups, remove them from the remaining groups to be checked. This saves on loop time and also removes the need to do `np.unique` on the final list.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

Happy new year!

The current redundancy finder relies on a function `find_clusters`, which groups points in 3D space that are all within a specified distance of each other. It does this by constructing a graph with the points as vertices and edges defined whenever two vertices are within the specified distance. The graph structure is expressed as a set of "adjacency lists" -- A list of lists where the k-th list gives the indices of vertices connected to vertex k. The vertices in each adjacency list are called "neighbors" of k.

Redundant groups are identified as _maximal cliques_ -- subsets of vertices such that each vertex in the subset is connected to each other, and no other vertices can be added while maintaining this "fully-connected" property. Apparently the task of identifying all maximal cliques in a graph if a well-known NP-complete problem. The existing approach is to loop over the neighbors (y) of each vertex (x) and identify when a x's adjacency list is a subset of y's. In that case, it means y is connected to all of the same points as x, and so it gets included in a group.

There is one other rule for identifying redundant groups that should be in place. In addition to being a maximal clique, a group of redundant baselines should also be disconnected from the rest of the graph such that no vertex is connected to a vertex outside the clique. I'll refer to these as "isolated cliques."

In practice, this means we can identify redundant groups as sets of vertices that all have the same adjacency list (if each vertex is considered adjacent to itself). Rather than looking at subsets then, we can look at equivalence, which can be a lot faster if we use hashing:
```
In [3]: s2 = frozenset(range(1000))

In [4]: s3 = frozenset(range(1000))

In [5]: %timeit s2 == s3
17.5 µs ± 308 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [6]: %timeit s2.__hash__() == s3.__hash__()
221 ns ± 1.71 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

Under the described changes, only isolated cliques will be counted as groups of redundant baselines, so we never end up with a baseline in multiple redundant groups. Such "ambiguous baselines" are detected when there are adjacency lists that are not in the final list of isolated cliques. Tolerance should always be chosen to produce a graph made up only of isolated cliques.

<!--- If it fixes an open issue, please link to the issue here. If this PR closes an issue, put the word 'closes' before the issue link to auto-close the issue when the PR is merged. -->


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Documentation change (documentation changes only)
- [ ] Version change
- [ ] Build or continuous integration change


## Checklist:
<!--- You may remove the checklists that don't apply to your change type(s) or just leave them empty -->
<!--- Go over all the following points, and replace the space with an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [contribution guide](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/blob/main/.github/CONTRIBUTING.md).
- [x] My code follows the code style of this project.

Bug fix checklist:
- [ ] My fix includes a new test that breaks as a result of the bug (if possible).
- [x] All new and existing tests pass.
- [x] I have updated the [CHANGELOG](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/blob/main/CHANGELOG.md).