@aelanman 2021-01-13
@bhazelton Thanks for the review! I've responded to each comment.